let filedeleting run in own thread. DocumentFile takes up to 10sec to delete one file.... wtf did google think about. Looks like best approach is to create a own FileSystemHandler from that all fileoperations are done async

+++ abstractcamera/src/main/java/com/troop/freedcam/utils/FileUtils.java
@@ -59,29 +59,29 @@ public class FileUtils
 
     public static boolean delteDocumentFile(File file)
     {
-        DocumentFile sdDir = FileUtils.getExternalSdDocumentFile();
-        String baseS = sdDir.getName();
-        String fileFolder = file.getAbsolutePath();
-        String[] split =  fileFolder.split("/");
-        DocumentFile tmpdir = null;
-        boolean append = false;
-        for (int t = 0; t < split.length; t++)
-        {
-            if (split[t].equals(baseS) ||append)
-            {
-                if (!append)
-                {
-                    append = true;
-                    tmpdir = sdDir;
-                }
-                else
-                {
-                    tmpdir = tmpdir.findFile(split[t]);
+        if (file.delete() == false) {
+            DocumentFile sdDir = FileUtils.getExternalSdDocumentFile();
+            String baseS = sdDir.getName();
+            String fileFolder = file.getAbsolutePath();
+            String[] split = fileFolder.split("/");
+            DocumentFile tmpdir = null;
+            boolean append = false;
+            for (int t = 0; t < split.length; t++) {
+                if (split[t].equals(baseS) || append) {
+                    if (!append) {
+                        append = true;
+                        tmpdir = sdDir;
+                    } else {
+                        tmpdir = tmpdir.findFile(split[t]);
+                    }
                 }
             }
+            boolean d = tmpdir.delete();
+            Logger.d("delteDocumentFile", "file delted:" + d);
+            return d;
         }
-        boolean d = tmpdir.delete();
-        Logger.d("delteDocumentFile", "file delted:" + d);
-        return d;
+        return true;
     }
+
+
 }
